aem65-an-abbottstore-apps/core/src/main/java/com/abbott/aem/an/abbottstore/models/Comparison.java
package com.abbott.aem.an.abbottstore.models;

import org.apache.sling.api.resource.Resource;
import org.apache.sling.models.annotations.DefaultInjectionStrategy;
import org.apache.sling.models.annotations.Model;
import org.apache.sling.models.annotations.injectorspecific.ValueMapValue;

/**
 *
 * @author saikrishna.s
 * 
 *         Comparison
 * 
 *         Comparison is the SlingModel to hold the details of individual
 *         Comparison.
 * 
 *         Version Number: 1.0
 */
@Model(adaptables = Resource.class, defaultInjectionStrategy = DefaultInjectionStrategy.OPTIONAL)
public class Comparison {

	/** The image. */
	@ValueMapValue
	private String image;

	/** The alt text. */
	@ValueMapValue
	private String altText;

	/**
	 * Gets the image.
	 *
	 * @return the image
	 */
	public String getImage() {
		return image;
	}

	/**
	 * Gets the alt text.
	 *
	 * @return the alt text
	 */
	public String getAltText() {
		return altText;
	}
}
